requirements.txt
matplotlib>=3.8.0
pandas>=2.2.0
seaborn>=0.13.0
scikit-learn>=1.4.0
streamlit>=1.33.0
yfinance>=0.2.40
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.1
transformers>=4.41.0,<5.0.0
datasets>=2.19.0
torch>=2.2.0,<2.6.0
accelerate>=0.30.0,<1.0.0
